break out logic into separate components
